The delete functionality and adding of excipient ingredients has been added. Once all delete, create, update and read functionaily is added, mapping for manufacturer foregin keys will be next.

+++ src/main/java/com/devstone/formulatorx/dao/ActiveIngredientRepository.java
@@ -4,6 +4,6 @@ import com.devstone.formulatorx.model.ActiveIngredient;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface ActiveIngredientRepository extends CrudRepository<ActiveIngredient, Long> {
+public interface ActiveIngredientRepository extends CrudRepository<ActiveIngredient, Integer> {
 
 }
