+++ src/main/java/com/devstone/formulatorx/dao/ExcipientIngredientRepository.java
@@ -4,6 +4,6 @@ import com.devstone.formulatorx.model.ExcipientIngredient;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface ExcipientIngredientRepository extends CrudRepository<ExcipientIngredient, Long> {
+public interface ExcipientIngredientRepository extends CrudRepository<ExcipientIngredient,Integer> {
 
 }
